change unsignedInteger

+++ database/migrations/2022_01_13_000001_update_forms_table.php
@@ -14,7 +14,7 @@ class UpdateFormsTable extends Migration
     public function up()
     {
         Schema::table('forms', function (Blueprint $table) {
-            $table->unsignedInteger('branch_id');
+            $table->unsignedBigInteger('branch_id');
             $table->foreign('branch_id')
                 ->references('id')
                 ->on('branches')
